# spring-boot-subscriptionservice/SubscriptionServiceSwagger.json
{"swagger":"2.0","info":{"description":"\"Spring Boot REST API for Subscription Service\"","version":"1.0.0","title":"Spring Boot REST API","license":{"name":"Apache License Version 2.0","url":"https://www.apache.org/licenses/LICENSE-2.0"}},"host":"localhost:8083","basePath":"/","tags":[{"name":"subscription-controller","description":"Subscription Service"}],"paths":{"/subscription/add":{"post":{"tags":["subscription-controller"],"summary":"Register Subscription","operationId":"saveSubscriptionUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"subscription","description":"subscription","required":true,"schema":{"$ref":"#/definitions/Subscription"}}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/subscription/list":{"get":{"tags":["subscription-controller"],"summary":"View a list of available subscriptions","operationId":"listUsingGET","produces":["application/json"],"responses":{"200":{"description":"Successfully retrieved list","schema":{"$ref":"#/definitions/Iterable"}},"401":{"description":"You are not authorized to view the resource"},"403":{"description":"Accessing the resource you were trying to reach is forbidden"},"404":{"description":"The resource you were trying to reach is not found"}},"deprecated":false}},"/subscription/show/{subscriptionId}":{"get":{"tags":["subscription-controller"],"summary":"Search a subscrition with an ID","operationId":"showSubscriptionUsingGET","produces":["application/json"],"parameters":[{"name":"subscriptionId","in":"path","description":"subscriptionId","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Subscription"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}}},"definitions":{"Iterable":{"type":"object","title":"Iterable"},"Iterable«Subscription»":{"type":"object","title":"Iterable«Subscription»"},"Subscription":{"type":"object","properties":{"dateOfBirth":{"type":"string","format":"date-time","description":"The date Of Birth"},"email":{"type":"string","description":"The email from customer"},"firstName":{"type":"string","description":"The auto-generated first name"},"flagConsent":{"type":"boolean","description":"The flag Consent"},"gender":{"type":"string","description":"The gender"},"newsletterId":{"type":"string","description":"The newsletter Id"},"subscriptionId":{"type":"integer","format":"int32","description":"The subscription id"}},"title":"Subscription"}}}